+++ src/main/ts/ephox/alloy/construct/CustomDefinition.ts
@@ -48,7 +48,7 @@ const toInfo = (spec: SimpleOrSketchSpec): Result<CustomDetail<any>, any> => {
         'alloy.focus': [ 'alloy.base.behaviour', 'focusing', 'keying' ],
         'alloy.system.init': [ 'alloy.base.behaviour', 'disabling', 'toggling', 'representing' ],
         'input': [ 'alloy.base.behaviour', 'representing', 'streaming', 'invalidating' ],
-        'alloy.system.detached': [ 'alloy.base.behaviour', 'representing' ],
+        'alloy.system.detached': [ 'alloy.base.behaviour', 'representing', 'item-events', 'tooltipping' ],
         'mousedown': [ 'focusing', 'alloy.base.behaviour', 'item-type-events' ],
         'mouseover': [ 'item-type-events', 'tooltipping' ]
       }),